fix(lens): minor polish from the lens review
- lowy F13 — `attentionNav.ts` held an app-global mutable jump registry AND a
  pure generic list helper, so a consumer wanting the cycling arithmetic had
  to import from the attention navigation registry to get it. `nextAfter`
  moved to a neutral leaf (`ui/nextAfter.ts`) with its own test; the registry
  keeps one job.
- hickey F12 — six comments, one of them written by this diff, sent the next
  reader to modules the PR deleted: `padi/surface.ts` justified the NEW
  `lingerIds` field by appealing to `pipIsActive`, a symbol that exists
  nowhere, and `Terminal.tsx`, `schema.ts`, `terminalActivityTracker.ts` and
  `effective-finish.mdx` still pointed at `useTerminalActivity`. All repointed
  at `attentionActive` / `useAttentionFacts`.
- Atlas `agent-attention` updated for the behaviour these commits changed —
  the class is derived once, paint is that partition renamed, a capsule jumps
  over exactly what it counted, and the rust leg is ACTIVE, not working. Dist
  regenerated with `just atlas::build`.

Raised by the lowy ∥ hickey lens review. Not pushed or merged.
