+++ packages/client/src/attention/attentionNav.test.ts
@@ -5,7 +5,7 @@
  *  and never reaches the rest. */
 
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { jumpToAsking, nextAfter, registerAttentionJump } from "./attentionNav";
+import { jumpToAsking, registerAttentionJump } from "./attentionNav";
 
 let dispose: (() => void) | null = null;
 afterEach(() => {
@@ -38,25 +38,5 @@ describe("registerAttentionJump", () => {
     dispose = registerAttentionJump(second);
     jumpToAsking("host-b");
     expect(second).toHaveBeenCalledWith("host-b");
-  });
-});
-
-describe("nextAfter", () => {
-  it("walks past the one you are already on, so repeat clicks reach them all", () => {
-    expect(nextAfter(["a", "b", "c"], "a")).toBe("b");
-    expect(nextAfter(["a", "b", "c"], "b")).toBe("c");
-  });
-
-  it("wraps at the end", () => {
-    expect(nextAfter(["a", "b"], "b")).toBe("a");
-  });
-
-  it("starts at the beginning when nothing is active, or the active one is elsewhere", () => {
-    expect(nextAfter(["a", "b"], null)).toBe("a");
-    expect(nextAfter(["a", "b"], "z")).toBe("a");
-  });
-
-  it("has nowhere to go in an empty set", () => {
-    expect(nextAfter([], null)).toBeUndefined();
   });
 });
