+++ packages/terminal-vocab/src/schema.ts
@@ -476,7 +476,7 @@ export type ProcessRss = z.infer<typeof ProcessRssSchema>;
 /** Output quiet-period before a terminal reads as static again — the ONE cadence
  *  the "is this terminal moving bytes right now" signal breathes at on BOTH sides:
  *  padi's `activity` stream (the padi-tui `●`) and the client's
- *  `useTerminalActivity` (the browser green dot). It is a RAW byte-motion signal
+ *  `useAttentionFacts` mirror (the browser green dot). It is a RAW byte-motion signal
  *  with a ~1s trailing window — a stream with sub-second gaps (compiles, `tail -f`)
  *  stays lit, one that pauses longer blinks off then back on when it resumes (so an
  *  agent that pauses >1s between thinking and emitting tokens flickers, by design).
